Break up setsockopt support and its TCP options.
This makes sockopt available to OS X and very likely to other BSD systems, more
especifically, OS X does not have TCP_KEEPIDLE, but has all the others.

If we can't detect the TCP option, we define the value to -1, then we fake a
ENOPROTOOPT if the user passed an option defined to -1.

The risk, although virtually impossible, is that there a unix system actually
using -1 to the value of those tcp options.

 checking VMSPLICE............failed
-checking SOCKOPT.............failed
+checking SOCKOPT.............ok
+checking TCP_KEEPCNT.........ok
+checking TCP_KEEPIDLE........failed
+checking TCP_KEEPINTVL.......ok
 checking POLL................ok